chore(gradle): extract system gradle support check to testutil
