+++ tsconfig.app.json
@@ -14,6 +14,7 @@
     "./dist",
     "**/__mocks__/**",
     "**/__fixtures__/**",
+    "**/__testutil__/**",
     "**/*.spec.ts",
     "./test"
   ]
